Move compatibility PByte declaration to implementation seciton
PByte declaration was added because older compilers didn't support pointer arithematic on PByte. So Mike decalred PByte as PAnsiChar - overriding the declaration in System.pas

By the time Compiler 20 came along, it *did* support arithmetic on PByte.

But in the meantime, if you're in Compiler before 20, then we'll be polluting into the global namespace a declaration of PByte that is incompatible with everything and everyone else. keep this declaration of PByte private to VirtualTrees.pas implementation section.